In term_switch move the label creation code down to where the label is used.
Remove redundant prototypes for temu_terminal_insert_text and temu_terminal_get_selection_text.

In temu_parse_config init file to NULL, check for all of the regex cases to succeed creation, and make sure that all of the regex cases are properly freed.

In main, free a fair bit on exit.  It's not even close to everything, but it's a start for using valgrind to make sure that we have no memory leaks.