Define the maximum number of windows.
This is the start of support for multiple windows in a single process.

The intent is that you can move a tab to a given window, and switching
to a tab belonging to another window will focus that window.

It is intended to be trivial to create and destroy windows.