Rewrite term_died to be multi-window aware.
When a terminal dies, we have to remove it from the correct notebook,
which means from the correct window.

It turns out that the user_data is not properly passed, which means that
we need to loop through all active terminals and use that index to find
the correct window.

terms.active[n] is no longer set to NULL, as term_destroyed handles
that.

And finally, prune_windows is called, so we can get rid of a window
after the last terminal in it exits.