Move notebook adding to new term_set_window.
The new function term_set_window handles the details of assigning a
given terminal to a given window.

This includes pieces that formerly lived in the new term section of
term_switch, sugn as a adding the terminal widget to the notebook.

It also includes things like making sure that the terminal isn't already
attached to another window.