Pass the term index to term_button_event.
The button press event is responsible for bringing up the menu on
alternate-button click.

Because the menu is window specific, and because the button press event
is term specific, we need to give it the index to the term, not a
pointer or index for the window.

This is due to the fact that the term can move between windows during
it's lifetime.

Also, setup the event _after_ setting up the active array, not before,
to avoid race conditions.