Rewrite the window title setting code.
Make it work at all, as we were not using the correct variable to find
the window.

Add a window ID (starts at 1) as a prefix if we have multiple windows.

Drop the window pointer argument to temu_window_title_change, we have to
look it up regardless in some cases.

Only the one asprintf call now, instead of it being in a conditional.