Fix the order in term_set_window.
The order in which events occur when we are setting a term to a given
window matters a bit.

We want to make the new window (if needed) as the first thing, among
other things this means that the window title changes that occur to the
old window as part of the removal of the term know that we have multiple
windows.

We also want to set the active window for the terminal right after that
for very similar reasons.