Add a menu option to bring up list of terms.
The new 'Show terminals' menu option brings up the notebook popup menu,
with a list of all active terminals.

This is in tab order, with the tab menu labels, which thanks to previous
commits means that it is in a sane order, and that the labels should
match the window labels.