Massive oversized GTK4 support.
This really should have been split into a dozen commits, but, well,
here's GTK4 support.

It seems to be largely functional, but it's not my daily driver yet, and
I don't expect it to even compile on Darwin at the moment.

(For one thing, brew isn't shipping a libvte compiled against gkt4 yet.)