Add dupstr, because strdupa isn't standard.
Darn MacOS.

Note, this is reasonably dangerous to use, as if there are more than 16
calls in sequence, we will start overwriting the older strings.

That's not a big deal if you only use it 16 or fewer calls at a time,
but...

At some point, I'll rewrite this to use alloca.