Try to replace geometry hints for window resizing.
The extra width/height pieces should probably become configuration
settings at some point.

This is definitely not as good as geometry hints, on the other hand,
Wayland decided not to have geometry hints, so they got dropped from
GTK4, and I definitely can't be bothered to try and manually implement
them for MacOS.

This will definitely be wrong in cases where there are things that count
towards the size of the window which are not the terminal area.

That _might_ include the top window bar, and definitely includes the tab
bar if that's enabled.

I'll put some thought into seeing if it's possible to get the size of
the tab bar at least at somepoint.