Significant rework of size management.
We had it working pretty well...  On MacOS.

It turns out that there are some significant differences between what
you have to do there, and what you have to do on Linux under X11.

I suspect that Wayland will have it's own quirks.

But so far this seems to be working pretty well, and it's definitely
more robust than it used to be.