Handle a failure to spawn.
In the event that we fail to spawn the program specified for this tab,
we need to correctly prune everything.

The easiest way to do that is to call term_died.

We use an exit status of -1 as a more or less random value that we
should never see otherwise.