Use the setup_context_menu signal for some menus.
This means that we no longer need to handle mouse clicks bringing up the
context menu.

However, I don't currently see any way to trigger this ourselves, so we
can't use this mechanism for bringing the menu up via a key binding.

This is, increasingly, a problem, as there are now several hard coded
key bindings in VTE.

Sure, they are for quite useful things, but it would be lovely to have
options.