Handle bindings involving shift.
Key bindings that involve shift are a problem, because we may get the
upper case keyval (C instead of c), and yet modern gtk_accelerator_parse
will always convert the keyval to lowercase.

We could do something outright absurd, like using gtk_accelerator_name
followed by gtk_accelerator_parse to normalize the key press, but the
performance impact of that makes it feel like a bad plan.

So if it's an alpha character, make a lower case version, and check the
binding against that as well.

This isn't necessarily perfect, but it seems to be working.