Rework the terminal list menu
It is no longer the GTK notebook popup, which hasn't worked reliably
with GTK4.

Now it's a dedicated menu.

Each entry has the title that shows up as the window title when that
table is active, and that title now includes the pts of the terminal.

This should make it _much_ easier to find things with ps, as it can list
the pts a program is running under.