Rework the menu in a couple of ways.
First, we're not using client side decorations for the header bar.

And the menu is accessible from the top right button on said header bar.

That menu is also now the one which is opened when you open the menu
with a key binding.

And finally, the logic for getting the location of the cursor when the
menu is opened by a key press has been rewritten, as it wasn't getting
the location right previously.